move the players/watchers computed properties to the game model
