+++ app/models/game.js
@@ -8,5 +8,38 @@ export default DS.Model.extend({
   gameParticipants: DS.hasMany('gameParticipant', {embedded: true}),
 
   /** @property {String} One of 'preparing', 'in-play', 'ended'. */
-  status: DS.attr('string', {defaultValue: 'preparing'})
+  status: DS.attr('string', {defaultValue: 'preparing'}),
+
+  /** @property {Array<GameParticipant>} People playing the game. */
+  players: function () {
+    var gameParticipants = this.get('gameParticipants');
+    var players = gameParticipants.filterBy('isPlaying');
+    while (players.length < 2) {
+      players.pushObject({
+        user: {
+          avatarUrl: 'http://www.gravatar.com/avatar/?s=256&default=mm',
+          username: '???'
+        }
+      });
+    }
+    return players;
+  }.property(
+    'gameParticipants.@each.user.username',
+    'gameParticipants.@each.deckName'),
+
+  /** @property {GameParticipant} The first player. */
+  playerOne: function () {
+    return this.get('players').objectAt(0);
+  }.property('players.@each'),
+
+  /** @property {GameParticipant} The second player. */
+  playerTwo: function () {
+    return this.get('players').objectAt(1);
+  }.property('players.@each'),
+
+  /** @property {Array<GameParticipant>} People watching the game. */
+  watchers: function () {
+    var gameParticipants = this.get('gameParticipants');
+    return gameParticipants.rejectBy('isPlaying');
+  }.property('gameParticipants.@each.isPlaying')
 });
